refactor(lane_departure_checker): move member functions to util functions
