+++ control/autoware_lane_departure_checker/src/lane_departure_checker_node/utils.cpp
@@ -19,6 +19,8 @@
 
 #include <boost/geometry.hpp>
 
+#include <lanelet2_core/geometry/Polygon.h>
+
 #include <vector>
 
 namespace
@@ -163,6 +165,64 @@ std::vector<LinearRing2d> createVehicleFootprints(
   }
 
   return vehicle_footprints;
+}
+
+lanelet::ConstLanelets getCandidateLanelets(
+  const lanelet::ConstLanelets & route_lanelets,
+  const std::vector<LinearRing2d> & vehicle_footprints)
+{
+  lanelet::ConstLanelets candidate_lanelets;
+
+  // Find lanes within the convex hull of footprints
+  const auto footprint_hull = createHullFromFootprints(vehicle_footprints);
+
+  for (const auto & route_lanelet : route_lanelets) {
+    const auto poly = route_lanelet.polygon2d().basicPolygon();
+    if (!boost::geometry::disjoint(poly, footprint_hull)) {
+      candidate_lanelets.push_back(route_lanelet);
+    }
+  }
+
+  return candidate_lanelets;
+}
+
+LinearRing2d createHullFromFootprints(const std::vector<LinearRing2d> & footprints)
+{
+  MultiPoint2d combined;
+  for (const auto & footprint : footprints) {
+    for (const auto & p : footprint) {
+      combined.push_back(p);
+    }
+  }
+
+  LinearRing2d hull;
+  boost::geometry::convex_hull(combined, hull);
+
+  return hull;
+}
+
+std::vector<LinearRing2d> createVehiclePassingAreas(
+  const std::vector<LinearRing2d> & vehicle_footprints)
+{
+  if (vehicle_footprints.empty()) {
+    return std::vector<LinearRing2d>();
+  }
+
+  std::vector<LinearRing2d> areas;
+  areas.reserve(vehicle_footprints.size() == 1 ? 1 : vehicle_footprints.size() - 1);
+
+  if (vehicle_footprints.size() == 1) {
+    areas.push_back(vehicle_footprints.front());
+    return areas;
+  }
+
+  for (size_t i = 0; i < vehicle_footprints.size() - 1; ++i) {
+    const auto & footprint1 = vehicle_footprints.at(i);
+    const auto & footprint2 = vehicle_footprints.at(i + 1);
+    areas.push_back(createHullFromFootprints({footprint1, footprint2}));
+  }
+
+  return areas;
 }
 
 PoseDeviation calcTrajectoryDeviation(
